requirements.txt
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.3
seaborn==0.12.0
scikit-learn==1.2.0
pingouin==0.5.3
shap==0.41.0
statsmodels==0.13.5
scipy==1.9.3
PyYAML==6.0
